requirements.txt
pygame~=2.6.1
PyQt6~=6.9.1
mutagen~=1.47.0
g2p-en~=2.1.0
pydub~=0.25.1
openai-whisper~=20240930
nltk~=3.9.1